pyproject.toml
[tool.poetry]
name = "csu"
version = "0.1.0"
description = ""
authors = ["Jeff Stein"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
rich = "^12.6.0"
aioauth-fastapi = "^0.1.1"
fastapi-utils = "^0.2.1"
playwright = "^1.27.1"
uvicorn = {extras = ["standard"], version = "^0.19.0"}
aiofiles = "^22.1.0"

[tool.poetry.group.dev.dependencies]
black = "^22.10.0"
mypy = "^0.982"
flake8 = "^5.0.4"
pre-commit = "^2.20.0"
types-aiofiles = "^22.1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.isort]
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true
line_length = 120
skip_gitignore = true
# you can skip files as below
#skip_glob = docs/conf.py


[tool.mypy]
# TODO: Enable mypy plugin after pydantic release supporting toml file
disallow_untyped_calls = true
disallow_untyped_defs = true
disallow_incomplete_defs = true

[tool.flake8]
ignore = "E501"
